Migrate sendSignal arg types into types/
