+++ packages/browser/src/methods/sendSignal.ts
@@ -1,5 +1,16 @@
-import type { Base64URLString, PublicKeyCredentialFuture } from '../types/index.ts';
+import type {
+  PublicKeyCredentialFuture,
+  SendSignalAllAcceptedCredentialsOpts,
+  SendSignalCurrentUserDetailsOpts,
+  SendSignalUnknownCredentialOpts,
+} from '../types/index.ts';
 import { identifySignalError } from '../helpers/identifySignalError.ts';
+
+export type {
+  SendSignalAllAcceptedCredentialsOpts,
+  SendSignalCurrentUserDetailsOpts,
+  SendSignalUnknownCredentialOpts,
+} from '../types/index.ts';
 
 /**
  * Broadcast a passkey state change on the server to the browser to enlist the browser's help
@@ -108,69 +119,3 @@ async function _callSignalCurrentUserDetails(opts: SendSignalCurrentUserDetailsO
 
   return undefined;
 }
-
-/**
- * A signal that communicates that the credential that the user just tried to register, or to
- * authenticate with, was not one that the Relying Party recognizes. The authenticator responsible
- * for the credential can hide or delete the credential so that the user does not see it in the
- * future as an option to sign in with.
- *
- * It is a good idea for a Relying Party to send this signal immediately after the use of an
- * unrecognized credential. For example, after rejecting the output from `startRegistration()` due
- * to unsatisfied RP-specific authenticator registration policy; or after rejecting the output from
- * `startAuthentication()` because the user deleted the passkey from their RP-specific user
- * settings.
- *
- * See https://w3c.github.io/webauthn/#sctn-signalUnknownCredential for more info.
- */
-export type SendSignalUnknownCredentialOpts = {
-  signalName: 'unknownCredential';
-  /** The same value used for `rpID` when calling \@simplewebauthn/server's `generateRegistrationOptions()` */
-  rpID: string;
-  /** The credential ID that the Relying Party didn't recognize for use */
-  credentialID: Base64URLString;
-};
-
-/**
- * A signal that communicates the current list of passkeys the Relying Party will recognize for use
- * by the **authenticated** user on the next login. Authenticators that have a passkey for
- * (rpId + userId), but the passkey ID is not found in allAcceptedCredentialIds, may choose to hide
- * or delete the passkey because it will not be accepted for use by the Relying Party.
- *
- * It is a good idea for a Relying Party to periodically send this signal, for example after every
- * successful authentication.
- *
- * See https://w3c.github.io/webauthn/#sctn-signalAllAcceptedCredentials for more info.
- */
-export type SendSignalAllAcceptedCredentialsOpts = {
-  signalName: 'allAcceptedCredentials';
-  /** The same value used for `rpID` when calling \@simplewebauthn/server's `generateRegistrationOptions()` */
-  rpID: string;
-  /** The base64url-encoded value used for `userID` when calling \@simplewebauthn/server's `generateRegistrationOptions()` */
-  userID: Base64URLString;
-  /** An array of base64url-encoded credential IDs for all credentials the user may use to authenticate */
-  allAcceptedCredentialIDs: Base64URLString[];
-};
-
-/**
- * A signal that communicates a change in the **authenticated** user's name and/or display name.
- * This can help browsers and platforms display the most up-to-date information about the user
- * during a passkey authentication instead of always showing whatever value was set at the time of
- * registration.
- *
- * It is a good idea for a Relying Party to periodically send this signal, for example after every
- * successful authentication and immediately after the user name and/or display name is changed.
- *
- * See https://w3c.github.io/webauthn/#sctn-signalCurrentUserDetails for more info.
- */
-export type SendSignalCurrentUserDetailsOpts = {
-  signalName: 'currentUserDetails';
-  /** The same value used for `rpID` when calling \@simplewebauthn/server's `generateRegistrationOptions()` */
-  rpID: string;
-  /** The base64url-encoded value used for `userID` when calling \@simplewebauthn/server's `generateRegistrationOptions()` */
-  userID: Base64URLString;
-  /** The primary account name, like an email address, username, etc... */
-  userName: string;
-  /** An optional, longer user identifier, like a full name, account differentiator, etc... Defaults to `""` */
-  userDisplayName?: string;
-};
